(RE-4635) Move Razor to a vanagon project
This work translates https://github.com/puppetlabs/razor-server and
https://github.com/puppetlabs/razor-torquebox to a vanagon project. This
allows us to more easily control the version of JRuby and Torquebox
delivered with Razor Server and used to build Razor Server. This shift
also allows us to take advantage of the VMPooler to provide build hosts
rather than relying on the COWs and MOCKs maintained on our rpm and deb
builders. This opens Razor Server up to allow builds on more esoteric
platforms if we so decide.