(maint) Update Pathing
This commit updates the pathing for razor-server/razor-torquebox as it
is installed on the users' machine. This update brings FOSS razor-server
more in line with the layout used for pe-razor-servr and pe-razor-libs.
The goal is that this package may eventually replace both
pe-razor-server and pe-razor-libs so that we only have to deliver one
razor package for both our open source and enterprise customers.

This pathing update also brings razor up to date in terms of layout
requirements for Puppet Collection packages.