(maint) Build razor-server using puppetlabs repo
Prior to this, razor-vanagon was pulling the razor-server code from a
fork rather than the upstream repository. This was because the required
changes to that repo had yet to be merged in. Now that those changes
have been merged in, we need to switch over to the main source
repository.