(maint) Deal with old directories
This commit updates the postinst actions on upgrade. We need to make
sure the user doesn't try to make any changes to any old config files.
We've already migrated the old files to the new location so they can be
found. Now, we need to clean up anything the packaging system didn't
clean up.

We also need to maintain the old repo-store directory if it is still in
use. The debian packaging system decided that it needed to remove the
razor user. I have no idea why it does this. So, we ensure the razor
user exists in the new packaging, and we make sure the old repo-store
has the correct permissions. Otherwise, users may be using the old
repo-store directory and the service will break because the razor user
no longer has permission to modify that directory.