(maint) Modify version
Prior to this commit, razor-server was installing a file that simply had
version set to DEVELOPMENT. Previous packaging implementations had taken
this string and modified it to report the real package version, rather
than DEVELOPMENT. This was missing in the conversion from the packaging
repo to Vanagon. This commit fixes that.