(RE-8457) Update to ship to puppet5
This commit updates razor-server to ship to the new puppet5 repos. This
change also includes settings for our nonfinal repos. This means that we
can ship development/beta/nightly builds to our nightly repos.

When razor-server ships to the puppet5 repos, we need to ensure what we
ship works with the puppet5 ecosystem.