(maint) Change user shell to /sbin/nologin
This commit changes the project user's shell from '/bin/bash' to '/sbin/nologin'.
Previously, the use of `su` with the `-c` flag when running the razor-server
service required the service user to have a login shell (see
https://bugzilla.redhat.com/show_bug.cgi?id=641101). Now that we use `runuser`
with the `-s` flag to specify the shell, we can safely set up the service user
without a login shell.