(RE-10703) Remove 'project' from build_defaults
This commit removes the 'project' key from build_defaults.yaml. This project
name is used to determine which directory to ship packages to on builds (among
other things), which lead to both foss razor-server and pe-razor-server
packages living in the same directory. This caused hiccups with PE promotion
and foss shipping, so we will now set the project as an environment variable
at build time, so that these packages end up in separate directories.